requirements_advanced.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
PyPDF2==3.0.1
requests==2.31.0
python-dotenv==1.0.0
arxiv==2.1.0
feedparser==6.0.10
beautifulsoup4==4.12.2
lxml==4.9.3
pydantic==2.5.0
aiohttp==3.9.1
nltk==3.8.1
scikit-learn==1.3.2
numpy==1.24.3
